requirements.txt
audioread==3.0.1
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
cycler==0.12.1
decorator==5.2.1
evdev==1.9.2
filelock==3.13.1
fsspec==2024.6.1
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
librosa==0.8.1
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.4.3
mpmath==1.3.0
networkx==3.0
numba==0.58.1
numpy==1.22.4
packaging==25.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.23
pynput==1.7.5
pyparsing==3.1.4
pyserial==3.5
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
python-xlib==0.33
requests==2.32.4
resampy==0.4.3
scikit-learn==1.3.2
scipy==1.7.3
shiboken2==5.15.2.1
six==1.17.0
sounddevice==0.4.3
soundfile==0.13.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1+cpu
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
